исправление задач 1, 3, 5

+++ tasks/task_5.py
@@ -1,6 +1,6 @@
 a = 10
 b = 3
 
-result = float(a / b)
+result = a / b
 
 print("Результат", result)
